requirements.txt
fastapi
uvicorn[standard]
pytesseract
Pillow
pdf2image
faiss-cpu
sentence-transformers
openai
python-multipart
